compacting
There are others working on the same project, but
- this fix makes a good generalised 'compact' without breaking
any of the ather changes.
Feel free to discard them, if they do not work in the longer run.
(the more spesific layout handling the better).
Mine is something in between, dealing with the group of parents,
children etc. so maybe it is needed after all.
As of 0.26 i think the display is nice.