The whole compacting project - complete
It adjusts nodelength with shorter notes in a narrow window,
there are no overlaps between nextfriends and siblings.

Adjusts between compact and normal nicely.

If user increases padding or font size or swithes to a wider
font type for a particular node #type above the base font
(whitch all calculations now hinge on), there *will* be overlaps.

But for now, without time consuming measuring out the position
of each node, this has to be the best approach?